requirements.txt
attrs==20.3.0
beautifulsoup4==4.10.0
blinker==1.4
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.4
colorama==0.4.4
comtypes==1.1.11
cryptography==36.0.1
docopt==0.6.2
face-recognition-models==0.3.0
Flask==2.0.3
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
jellyfish==0.8.9
Jinja2==3.0.3
Js2Py==0.71
lib50==3.0.3
MarkupSafe==2.1.1
numpy==1.22.2
packaging==21.3
pexpect==4.8.0
Pillow==9.0.1
pipwin==0.5.1
ptyprocess==0.7.0
PyAudio @ file:///C:/Users/hp/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pycparser==2.21
pyjsparser==2.7.1
pyparsing==3.0.7
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==0.20.0
pyttsx3==2.90
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.36
submit50==3.1.0
termcolor==1.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validate-email==1.3
Werkzeug==2.0.3
WTForms==3.0.1
